fix: resolve code review comments
